+++ tests/generators/open-graph-locale-generator-test.php
@@ -10,7 +10,7 @@ namespace Yoast\WP\SEO\Tests\Generators;
 use Brain\Monkey;
 use Mockery;
 use Yoast\WP\SEO\Generators\Open_Graph_Locale_Generator;
-use Yoast\WP\SEO\Tests\Mocks\Meta_Tags_Context_Mock;
+use Yoast\WP\SEO\Tests\Doubles\Context\Meta_Tags_Context_Mock;
 use Yoast\WP\SEO\Tests\TestCase;
 
 /**
