+++ tests/presentations/indexable-static-posts-page-presentation/presentation-instance-builder.php
@@ -10,7 +10,7 @@ use Yoast\WP\SEO\Helpers\Post_Type_Helper;
 use Yoast\WP\SEO\Presentations\Indexable_Post_Type_Presentation;
 use Yoast\WP\SEO\Presentations\Indexable_Static_Posts_Page_Presentation;
 use Yoast\WP\SEO\Tests\Mocks\Indexable_Mock;
-use Yoast\WP\SEO\Tests\Mocks\Meta_Tags_Context_Mock;
+use Yoast\WP\SEO\Tests\Doubles\Context\Meta_Tags_Context_Mock;
 use Yoast\WP\SEO\Tests\Presentations\Presentation_Instance_Dependencies;
 
 /**
